refactor: fix files structure
