+++ apps/cowswap-frontend/src/modules/hooksStore/dapps/ClaimGnoHookApp/const.ts
@@ -1,0 +1,1 @@
+export const SBC_DEPOSIT_CONTRACT_ADDRESS = '0x0B98057eA310F4d31F2a452B414647007d1645d9'
