+++ apps/cowswap-frontend/src/modules/hooksStore/dapps/ClaimGnoHookApp/useSBCDepositContract.ts
@@ -4,13 +4,13 @@ import { useWalletInfo } from '@cowprotocol/wallet'
 
 import { useContract } from 'common/hooks/useContract'
 
-import { SBC_DEPOSIT_CONTRACT_ADDRESS } from '../const'
+import { SBC_DEPOSIT_CONTRACT_ADDRESS } from './const'
 
 export function useSBCDepositContract(): SBCDepositContract | null {
   const { chainId } = useWalletInfo()
   return useContract<SBCDepositContract>(
     chainId === SupportedChainId.GNOSIS_CHAIN ? SBC_DEPOSIT_CONTRACT_ADDRESS : undefined,
     SBCDepositContractAbi,
-    true
+    true,
   )
 }
